+++ CHANGELOG.md
@@ -3,6 +3,7 @@ DevOps Tool: Magento 2 Platform Support Changelog
 
 # 0.2.0 (Unreleased)
 - Updated PHP version requirement to 7.1
+- Updated config to work with DevopsToolAppOrchestration
 
 # 0.1.1
 - Added zf config-provider Composer setting